requirements.txt
asgiref==3.8.1
Django==5.1.3
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
openpyxl==3.1.5
PyJWT==2.10.0
sqlparse==0.5.2
tzdata==2024.2
